+++ src/components/Initiative/config.js
@@ -6,6 +6,21 @@ const config = {
       size: 'l',
     },
   },
+  error: {
+    variant: 'error',
+    icon: {
+      shape: 'notifications--error',
+      size: 'l',
+    },
+    close: {
+      variant: 'ghost',
+      label: 'Close',
+      icon: {
+        shape: 'ui--close',
+        size: 's',
+      },
+    },
+  },
 };
 
 export default config;
