+++ src/pages/sitemap.jsx
@@ -18,8 +18,8 @@ const Sitemap = ({ location }) => {
           <div className="ecl-row ecl-u-mt-l">
             <div className="ecl-col-sm-12 ecl-col-md-12">
               <h1 className="ecl-u-type-heading-1">Sitemap</h1>
-              <ul class="ecl-unordered-list">
-                <li class="ecl-unordered-list__item">
+              <ul className="ecl-unordered-list">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/`}
                     className="ecl-link ecl-link--standalone"
@@ -27,7 +27,7 @@ const Sitemap = ({ location }) => {
                     Home
                   </Link>
                 </li>
-                <li class="ecl-unordered-list__item">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/news`}
                     className="ecl-link ecl-link--standalone"
@@ -35,7 +35,7 @@ const Sitemap = ({ location }) => {
                     News
                   </Link>
                 </li>
-                <li class="ecl-unordered-list__item">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/how-to-start`}
                     className="ecl-link ecl-link--standalone"
@@ -43,7 +43,7 @@ const Sitemap = ({ location }) => {
                     How to start
                   </Link>
                 </li>
-                <li class="ecl-unordered-list__item">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/faq`}
                     className="ecl-link ecl-link--standalone"
@@ -51,7 +51,7 @@ const Sitemap = ({ location }) => {
                     FAQ
                   </Link>
                 </li>
-                <li class="ecl-unordered-list__item">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/contact`}
                     className="ecl-link ecl-link--standalone"
@@ -59,7 +59,7 @@ const Sitemap = ({ location }) => {
                     Contact
                   </Link>
                 </li>
-                <li class="ecl-unordered-list__item">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/sitemap`}
                     className="ecl-link ecl-link--standalone"
@@ -67,7 +67,7 @@ const Sitemap = ({ location }) => {
                     Sitemap
                   </Link>
                 </li>
-                <li class="ecl-unordered-list__item">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/communication-material`}
                     className="ecl-link ecl-link--standalone"
@@ -75,15 +75,15 @@ const Sitemap = ({ location }) => {
                     Communication material
                   </Link>
                 </li>
-                <li class="ecl-unordered-list__item">
+                <li className="ecl-unordered-list__item">
                   <Link
                     to={`/${currentLanguage}/how-it-works`}
                     className="ecl-link ecl-link--standalone"
                   >
                     How it works
                   </Link>
-                  <ul class="ecl-unordered-list">
-                    <li class="ecl-unordered-list__item">
+                  <ul className="ecl-unordered-list">
+                    <li className="ecl-unordered-list__item">
                       <Link
                         to={`/${currentLanguage}/how-it-works/regulatory-framework`}
                         className="ecl-link ecl-link--standalone"
@@ -91,7 +91,7 @@ const Sitemap = ({ location }) => {
                         Regulatory framework
                       </Link>
                     </li>
-                    <li class="ecl-unordered-list__item">
+                    <li className="ecl-unordered-list__item">
                       <Link
                         to={`/${currentLanguage}/how-it-works/implementation-national-level`}
                         className="ecl-link ecl-link--standalone"
@@ -99,7 +99,7 @@ const Sitemap = ({ location }) => {
                         Implementation national level
                       </Link>
                     </li>
-                    <li class="ecl-unordered-list__item">
+                    <li className="ecl-unordered-list__item">
                       <Link
                         to={`/${currentLanguage}/how-it-works/data-protection`}
                         className="ecl-link ecl-link--standalone"
@@ -107,7 +107,7 @@ const Sitemap = ({ location }) => {
                         Data protection
                       </Link>
                     </li>
-                    <li class="ecl-unordered-list__item">
+                    <li className="ecl-unordered-list__item">
                       <Link
                         to={`/${currentLanguage}/how-it-works/history`}
                         className="ecl-link ecl-link--standalone"
